ultimos cambios

+++ Kernel/include/exceptions.h
@@ -1,6 +1,6 @@
 #ifndef _exceptions
 #define _exceptions 
-void exceptionDispatcher(int exception,uint64_t rsp);
+void exceptionDispatcher(uint64_t exception,uint64_t rsp);
 void zero_division();
 void printAllRegisters(uint64_t rsp);
 int tryAndContinue();
